Update Chromium version 107.0.5304.17 to 93.0.4577.25

+++ content/public/android/java/src/org/chromium/content/browser/selection/SelectionPopupControllerImpl.java
@@ -1,9 +1,10 @@
-// Copyright 2015 The Chromium Authors
+// Copyright 2015 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package org.chromium.content.browser.selection;
 
+import android.annotation.TargetApi;
 import android.app.Activity;
 import android.app.SearchManager;
 import android.content.Context;
@@ -28,9 +29,9 @@ import android.view.textclassifier.SelectionEvent;
 import android.view.textclassifier.TextClassifier;
 
 import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 import androidx.annotation.VisibleForTesting;
 
+import org.chromium.base.BuildInfo;
 import org.chromium.base.Log;
 import org.chromium.base.PackageManagerUtils;
 import org.chromium.base.UserData;
@@ -39,9 +40,8 @@ import org.chromium.base.annotations.JNINamespace;
 import org.chromium.base.annotations.NativeMethods;
 import org.chromium.base.compat.ApiHelperForM;
 import org.chromium.base.metrics.RecordUserAction;
-import org.chromium.base.supplier.ObservableSupplier;
-import org.chromium.base.supplier.ObservableSupplierImpl;
 import org.chromium.content.R;
+import org.chromium.content.browser.ContentApiHelperForM;
 import org.chromium.content.browser.ContentClassFactory;
 import org.chromium.content.browser.GestureListenerManagerImpl;
 import org.chromium.content.browser.PopupController;
@@ -54,7 +54,6 @@ import org.chromium.content.browser.webcontents.WebContentsImpl.UserDataFactory;
 import org.chromium.content_public.browser.ActionModeCallbackHelper;
 import org.chromium.content_public.browser.ImeEventObserver;
 import org.chromium.content_public.browser.RenderFrameHost;
-import org.chromium.content_public.browser.SelectAroundCaretResult;
 import org.chromium.content_public.browser.SelectionClient;
 import org.chromium.content_public.browser.SelectionPopupController;
 import org.chromium.content_public.browser.WebContents;
@@ -119,7 +118,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     private Context mContext;
     private WindowAndroid mWindowAndroid;
     private WebContentsImpl mWebContents;
-    private ActionMode.Callback2 mCallback;
+    private ActionMode.Callback mCallback;
     private RenderFrameHost mRenderFrameHost;
     private long mNativeSelectionPopupController;
 
@@ -139,10 +138,6 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     @Nullable
     private View mView;
     private ActionMode mActionMode;
-
-    // Supplier of whether action bar is showing now.
-    private final ObservableSupplierImpl<Boolean> mIsActionBarShowingSupplier =
-            new ObservableSupplierImpl<>();
 
     // Bit field for mappings from menu item to a flag indicating it is allowed.
     private int mAllowedMenuItems;
@@ -340,7 +335,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     }
 
     @Override
-    public void setActionModeCallback(ActionMode.Callback2 callback) {
+    public void setActionModeCallback(ActionMode.Callback callback) {
         mCallback = callback;
     }
 
@@ -410,7 +405,8 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             boolean shouldSuggest, @MenuSourceType int sourceType,
             RenderFrameHost renderFrameHost) {
         int offsetBottom = bottom;
-        offsetBottom += handleHeight;
+        // Legacy action mode expects the selection rectangle not to include touch handle.
+        if (supportsFloatingActionMode()) offsetBottom += handleHeight;
         mSelectionRect.set(left, top, right, offsetBottom);
         mEditable = isEditable;
         mLastSelectedText = selectionText;
@@ -486,15 +482,23 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         destroyActionModeAndKeepSelection();
 
         assert mWebContents != null;
-        ActionMode actionMode = mView.startActionMode(mCallback, ActionMode.TYPE_FLOATING);
+        ActionMode actionMode = supportsFloatingActionMode() ? startFloatingActionMode()
+                                                             : mView.startActionMode(mCallback);
         if (actionMode != null) {
             // This is to work around an LGE email issue. See crbug.com/651706 for more details.
             LGEmailActionModeWorkaroundImpl.runIfNecessary(mContext, actionMode);
         }
-        setActionMode(actionMode);
+        mActionMode = actionMode;
         mUnselectAllOnDismiss = true;
 
         if (!isActionModeValid()) clearSelection();
+    }
+
+    private ActionMode startFloatingActionMode() {
+        assert mView != null;
+        assert Build.VERSION.SDK_INT >= Build.VERSION_CODES.M;
+        ActionMode actionMode = ContentApiHelperForM.startActionMode(mView, this, mCallback);
+        return actionMode;
     }
 
     private void dismissTextHandles() {
@@ -514,6 +518,10 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             return;
         }
 
+        if (!supportsFloatingActionMode() && !Clipboard.getInstance().canPaste()
+                && mNonSelectionCallback == null) {
+            return;
+        }
         destroyPastePopup();
         PastePopupMenu.PastePopupMenuDelegate delegate =
                 new PastePopupMenu.PastePopupMenuDelegate() {
@@ -551,8 +559,12 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
                 };
         Context windowContext = mWindowAndroid.getContext().get();
         if (windowContext == null) return;
-        mPastePopupMenu =
-                new FloatingPastePopupMenu(windowContext, mView, delegate, mNonSelectionCallback);
+        if (supportsFloatingActionMode()) {
+            mPastePopupMenu = new FloatingPastePopupMenu(
+                    windowContext, mView, delegate, mNonSelectionCallback);
+        } else {
+            mPastePopupMenu = new LegacyPastePopupMenu(windowContext, mView, delegate);
+        }
         showPastePopup();
     }
 
@@ -561,6 +573,11 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             mPastePopupMenu.show(getSelectionRectRelativeToContainingView());
         } catch (WindowManager.BadTokenException e) {
         }
+    }
+
+    @Override
+    public boolean supportsFloatingActionMode() {
+        return Build.VERSION.SDK_INT >= Build.VERSION_CODES.M;
     }
 
     // HideablePopup implementation
@@ -595,7 +612,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             mActionMode.finish();
 
             // Should be nulled out in case #onDestroyActionMode() is not invoked in response.
-            setActionMode(null);
+            mActionMode = null;
         }
     }
 
@@ -603,14 +620,16 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
      * @see ActionMode#invalidateContentRect()
      */
     public void invalidateContentRect() {
-        if (isActionModeValid()) ApiHelperForM.invalidateContentRectOnActionMode(mActionMode);
+        if (supportsFloatingActionMode() && isActionModeValid()) {
+            ApiHelperForM.invalidateContentRectOnActionMode(mActionMode);
+        }
     }
 
     // WindowEventObserver
 
     @Override
     public void onWindowFocusChanged(boolean gainFocus) {
-        if (isActionModeValid()) {
+        if (supportsFloatingActionMode() && isActionModeValid()) {
             ApiHelperForM.onWindowFocusChangedOnActionMode(mActionMode, gainFocus);
         }
     }
@@ -707,12 +726,15 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     }
 
     private boolean isFloatingActionMode() {
-        return isActionModeValid()
+        return supportsFloatingActionMode() && isActionModeValid()
                 && ApiHelperForM.getActionModeType(mActionMode) == ActionMode.TYPE_FLOATING;
     }
 
     private long getDefaultHideDuration() {
-        return ApiHelperForM.getDefaultActionModeHideDuration();
+        if (supportsFloatingActionMode()) {
+            return ApiHelperForM.getDefaultActionModeHideDuration();
+        }
+        return 2000;
     }
 
     // Default handlers for action mode callbacks.
@@ -763,7 +785,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         new MenuInflater(context).inflate(R.menu.select_action_menu, menu);
     }
 
-    @RequiresApi(Build.VERSION_CODES.O)
+    @TargetApi(Build.VERSION_CODES.O)
     public static void setPasteAsPlainTextMenuItemTitle(Menu menu) {
         MenuItem item = menu.findItem(R.id.select_action_menu_paste_as_plain_text);
         if (item == null) return;
@@ -823,7 +845,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             menu.removeItem(R.id.select_action_menu_web_search);
         }
 
-        if (isSelectionPassword() || !Clipboard.getInstance().canCopy()) {
+        if (isSelectionPassword()) {
             menu.removeItem(R.id.select_action_menu_copy);
             menu.removeItem(R.id.select_action_menu_cut);
         }
@@ -883,12 +905,12 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         }
     }
 
-    @RequiresApi(Build.VERSION_CODES.M)
+    @TargetApi(Build.VERSION_CODES.M)
     private static Intent createProcessTextIntent() {
         return new Intent().setAction(Intent.ACTION_PROCESS_TEXT).setType("text/plain");
     }
 
-    @RequiresApi(Build.VERSION_CODES.M)
+    @TargetApi(Build.VERSION_CODES.M)
     private Intent createProcessTextIntentForResolveInfo(ResolveInfo info) {
         boolean isReadOnly = !isFocusedNodeEditable();
         return createProcessTextIntent()
@@ -952,7 +974,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
 
     @Override
     public void onDestroyActionMode() {
-        setActionMode(null);
+        mActionMode = null;
         if (mUnselectAllOnDismiss) {
             clearSelection();
         }
@@ -1127,7 +1149,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         try {
             mWindowAndroid.showIntent(intent, new WindowAndroid.IntentCallback() {
                 @Override
-                public void onIntentCompleted(int resultCode, Intent data) {
+                public void onIntentCompleted(WindowAndroid window, int resultCode, Intent data) {
                     onReceivedProcessTextResult(resultCode, data);
                 }
             }, null);
@@ -1382,8 +1404,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     @CalledByNative
     /* package */ void onDragUpdate(@TouchSelectionDraggableType int type, float x, float y) {
         // If this is for longpress drag selector, we can only have mangifier on S and above.
-        if (type == TouchSelectionDraggableType.LONGPRESS
-                && Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
+        if (type == TouchSelectionDraggableType.LONGPRESS && !BuildInfo.isAtLeastS()) {
             return;
         }
 
@@ -1402,7 +1423,6 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         if (mWebContents == null || !isActionModeSupported()) return;
         mWebContents.collapseSelection();
         mClassificationResult = null;
-        mHasSelection = false;
     }
 
     private PopupController getPopupController() {
@@ -1480,19 +1500,9 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     }
 
     @CalledByNative
-    private void onSelectAroundCaretSuccess(int extendedStartAdjust, int extendedEndAdjust,
-            int wordStartAdjust, int wordEndAdjust) {
+    private void onSelectWordAroundCaretAck(boolean didSelect, int startAdjust, int endAdjust) {
         if (mSelectionClient != null) {
-            SelectAroundCaretResult result = new SelectAroundCaretResult(
-                    extendedStartAdjust, extendedEndAdjust, wordStartAdjust, wordEndAdjust);
-            mSelectionClient.selectAroundCaretAck(result);
-        }
-    }
-
-    @CalledByNative
-    private void onSelectAroundCaretFailure() {
-        if (mSelectionClient != null) {
-            mSelectionClient.selectAroundCaretAck(null);
+            mSelectionClient.selectWordAroundCaretAck(didSelect, startAdjust, endAdjust);
         }
     }
 
@@ -1531,15 +1541,11 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         return mLastSelectedText;
     }
 
-    private void setActionMode(ActionMode actionMode) {
-        mActionMode = actionMode;
-        mIsActionBarShowingSupplier.set(isSelectActionBarShowing());
-    }
-
     private boolean isShareAvailable() {
         Intent intent = new Intent(Intent.ACTION_SEND);
         intent.setType("text/plain");
-        return PackageManagerUtils.canResolveActivity(intent, PackageManager.MATCH_DEFAULT_ONLY);
+        return !PackageManagerUtils.queryIntentActivities(intent, PackageManager.MATCH_DEFAULT_ONLY)
+                        .isEmpty();
     }
 
     // The callback class that delivers the result from a SmartSelectionClient.
@@ -1596,11 +1602,6 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     @Override
     public boolean isSelectActionBarShowing() {
         return isActionModeValid();
-    }
-
-    @Override
-    public ObservableSupplier<Boolean> isSelectActionBarShowingSupplier() {
-        return mIsActionBarShowingSupplier;
     }
 
     @Override
